chore: moved helper for running e2e test
